Migrate 05-testing-react/my-app from Create React App to Vite
Agent-Logs-Url: https://github.com/simonachkar/mcda5512-software-testing/sessions/83719e81-87c9-48c6-9598-2c28d668a44e
