--- 05-testing-react/my-app/src/components/title/title.test.js
@@ -1,8 +1,0 @@
-import ReactDOM from "react-dom";
-import Title from "./title";
-
-test("renders <Title>", () => {
-  const div = document.createElement("div");
-  ReactDOM.render(<Title />, div);
-  expect(div.querySelector("h1").textContent).toBe("Hello World!");
-});
