+++ 05-testing-react/my-app/src/main.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+import ReactDOM from 'react-dom/client'
+import './index.css'
+import App from './App'
+
+const root = ReactDOM.createRoot(document.getElementById('root'))
+root.render(<App />)
